completed collision

+++ PaxLab/Common/src/main/java/dk/group6/common/data/entityparts/SpritePart.java
@@ -77,31 +77,31 @@ public class SpritePart implements EntityPart {
         return sprite;
     }
     
-    public float[] getSpriteLeftBottom() {
-        float[] coordinates = new float[2];
-        coordinates[0] = sprite.getX();
-        coordinates[1] = sprite.getY();
+    public int[] getSpriteLeftBottom() {
+        int[] coordinates = new int[2];
+        coordinates[0] = (int) sprite.getX();
+        coordinates[1] = (int) sprite.getY();
         return coordinates;
     }
     
-    public float[] getSpriteRightBottom() {
-        float[] coordinates = new float[2];
-        coordinates[0] = sprite.getX() + sprite.getWidth();
-        coordinates[1] = sprite.getY();
+    public int[] getSpriteRightBottom() {
+        int[] coordinates = new int[2];
+        coordinates[0] = (int) (sprite.getX() + sprite.getWidth());
+        coordinates[1] = (int) sprite.getY();
         return coordinates;
     }
     
-    public float[] getSpriteLeftTop() {
-        float[] coordinates = new float[2];
-        coordinates[0] = sprite.getX();
-        coordinates[1] = sprite.getY() + sprite.getHeight();
+    public int[] getSpriteLeftTop() {
+        int[] coordinates = new int[2];
+        coordinates[0] = (int) sprite.getX();
+        coordinates[1] = (int) (sprite.getY() + sprite.getHeight());
         return coordinates;
     }
     
-    public float[] getSpriteRightTop() {
-        float[] coordinates = new float[2];
-        coordinates[0] = sprite.getX() + sprite.getWidth();
-        coordinates[1] = sprite.getY() + sprite.getHeight();
+    public int[] getSpriteRightTop() {
+        int[] coordinates = new int[2];
+        coordinates[0] = (int) (sprite.getX() + sprite.getWidth());
+        coordinates[1] = (int) (sprite.getY() + sprite.getHeight()); 
         return coordinates;
     }
 }
